+++ pyts/transformation/weasel.py
@@ -134,7 +134,7 @@ class WEASEL(BaseEstimator, TransformerMixin):
         self : object
 
         """
-        X, y = check_X_y(X, y)
+        X, y = check_X_y(X, y, dtype='float64')
         check_classification_targets(y)
         n_samples, n_timestamps = X.shape
         window_sizes, window_steps = self._check_params(n_timestamps)
@@ -207,7 +207,7 @@ class WEASEL(BaseEstimator, TransformerMixin):
             check_is_fitted(self, ['_relevant_features_list', '_sfa_list',
                                    '_vectorizer_list', 'vocabulary_'])
 
-        X = check_array(X)
+        X = check_array(X, dtype='float64')
         n_samples, n_timestamps = X.shape
 
         X_features = coo_matrix((n_samples, 0), dtype=np.int64)
@@ -253,7 +253,7 @@ class WEASEL(BaseEstimator, TransformerMixin):
             Document-term matrix.
 
         """
-        X, y = check_X_y(X, y)
+        X, y = check_X_y(X, y, dtype='float64')
         check_classification_targets(y)
         n_samples, n_timestamps = X.shape
         window_sizes, window_steps = self._check_params(n_timestamps)
